+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:59513/jest?","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:59571/jest?","mongoDBName":"jest"}
